% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{KDE.int}
\alias{KDE.int}
\title{Kernel Density Estimation for Inter-event Times}
\usage{
KDE.int(intervals)
}
\arguments{
\item{intervals}{a vector of inter-event times in seconds}
}
\description{
This function receives
}
\examples{
intervals = c(1,5,2,3,15,2,354374,124135,352,234,14,5331)
KDE.int(intervals)

}

