% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotEvent}
\alias{PlotEvent}
\title{Timestamp plotting}
\usage{
PlotEvent(timestamps, ...)
}
\arguments{
\item{UNIX}{timestamps/epochs of an event series}
}
\description{
Plot timestamps with vertical lines
}

